+++ Arduino Files/Piglet/WebUI.cpp
@@ -256,7 +256,7 @@ static const char INDEX_HTML[] PROGMEM = R"HTML(
     <div class="logo">&#x1f437;</div>
     <div>
       <h1>Piglet Wardriver</h1>
-<p class="sub">ESP32 Wi-Fi Scanner &amp; Logger &mdash; v2.51</p>
+<p class="sub">ESP32 Wi-Fi &amp; BLE Scanner &mdash; v2.52</p>
     </div>
   </div>